refactor: 내부에서 쓰이는 스켈레톤 컴포넌트는 internal 폴더로 분리
