+++ src/components/skeleton/internal/skeletonContainer.ts
@@ -1,4 +1,4 @@
-import { createElementWithAttributes } from "../utils/createElementWithAttributes";
+import { createElementWithAttributes } from "../../utils/createElementWithAttributes";
 
 const skeletonContainer = (count: number) => {
   const $skeletonContainer = createElementWithAttributes({
